+++ Main/DirectX11_Starter_2015/DirectX11_Starter/FXAA.cpp
@@ -1,6 +1,6 @@
 #include "FXAA.h"
 
-FXAA::FXAA(float width, float height, ID3D11Device* _device, ID3D11DeviceContext * _deviceContext, ID3D11SamplerState* _sampler)
+FXAA::FXAA(int width, int height, ID3D11Device* _device, ID3D11DeviceContext * _deviceContext, ID3D11SamplerState* _sampler)
 {
 	windowWidth = width;
 	windowHeight = height;
