+++ Main/DirectX11_Starter_2015/DirectX11_Starter/Tint.h
@@ -5,7 +5,7 @@
 class Tint : public PostProcessBase
 {
 public:
-	Tint(float width, float height, ID3D11Device* _device, ID3D11DeviceContext * _deviceContext, ID3D11SamplerState* _sampler, ID3D11RenderTargetView* _rtv);
+	Tint(int width, int height, ID3D11Device* _device, ID3D11DeviceContext * _deviceContext, ID3D11SamplerState* _sampler, ID3D11RenderTargetView* _rtv);
 	~Tint();
 	virtual ID3D11ShaderResourceView* draw(ID3D11ShaderResourceView* ppSRV);
 private:
