+++ Main/DirectX11_Starter_2015/DirectX11_Starter/ppTest.h
@@ -6,7 +6,7 @@
 class ppTest : public PostProcessBase
 {
 public:
-	ppTest(float width, float height, ID3D11Device* _device, ID3D11DeviceContext * _deviceContext, ID3D11SamplerState* _sampler, ID3D11RenderTargetView* _rtv);
+	ppTest(int width, int height, ID3D11Device* _device, ID3D11DeviceContext * _deviceContext, ID3D11SamplerState* _sampler, ID3D11RenderTargetView* _rtv);
 	virtual ID3D11ShaderResourceView* draw(ID3D11ShaderResourceView* ppSRV);
 	~ppTest();
 private:
